+++ src/serde_multipart/unserializers/input_file.rs
@@ -62,7 +62,9 @@ impl Serializer for InputFileUnserializer {
         // TODO
         match variant {
             "File" => Ok(InputFile::File(value.serialize(StringUnserializer)?.into())),
-            "Url" => Ok(InputFile::Url(value.serialize(StringUnserializer)?)),
+            "Url" => Ok(InputFile::Url(
+                reqwest::Url::parse(&value.serialize(StringUnserializer)?).unwrap(),
+            )),
             "FileId" => Ok(InputFile::FileId(value.serialize(StringUnserializer)?)),
             name => Err(UnserializerError::UnexpectedVariant {
                 name,