+++ src/types/inline_query_result_cached_voice.rs
@@ -96,7 +96,7 @@ impl InlineQueryResultCachedVoice {
         self
     }
 
-    pub fn parse_mode<S>(mut self, val: ParseMode) -> Self {
+    pub fn parse_mode(mut self, val: ParseMode) -> Self {
         self.parse_mode = Some(val);
         self
     }