+++ src/types/inline_query_result_venue.rs
@@ -52,7 +52,7 @@ pub struct InlineQueryResultVenue {
     pub input_message_content: Option<InputMessageContent>,
 
     /// Url of the thumbnail for the result.
-    pub thumb_url: Option<String>,
+    pub thumb_url: Option<reqwest::Url>,
 
     /// Thumbnail width.
     pub thumb_width: Option<i32>,
@@ -162,11 +162,8 @@ impl InlineQueryResultVenue {
         self
     }
 
-    pub fn thumb_url<S>(mut self, val: S) -> Self
-    where
-        S: Into<String>,
-    {
-        self.thumb_url = Some(val.into());
+    pub fn thumb_url(mut self, val: reqwest::Url) -> Self {
+        self.thumb_url = Some(val);
         self
     }